Start using BeautifulHugo theme
This makes some customization to the BeautifulHugo theme by adding
an exact copy of footer, header and nav in the `layouts` directory
that basically is the same with some personal modifications. This
will cause new improvements in the theme in those specific files
not to be included or they can with a manual process.

Also part of the customization is modifying the original translat-
ions for specific parts of the template. It does not removes
the original author name but removes the Jerkyll text that I see
unnecessary.

As part of this update it adds support to English and Spanish
so new posts can exist in one or in two different languages.